requirements.txt
streamlit
pillow
numpy
tensorflow
pandas
